refactor(mp-8b1b): simplify-pass cleanup on the kachinuki PR
- Extract applyKachinukiMerge so the locked and tx scoring paths share
  one kachinuki merge guard instead of two verbatim copies.
- kachinuki_export.go reuses bracketMatchToTeamResult for both bracket
  projections instead of hand-rolled MatchResult literals.
- TeamScoreEditorModal computes kachinukiVisiblePositions once per
  render (was twice with identical inputs) and drops a dead alias.
- resolveBoutSideName moves to lineup_resolver.jsx; BoutSubRow and the
  streaming overlay now call the shared, tested helper instead of
  inlining the same name-priority logic.
- enrichPoolMatchWithComp stamps teamMatchType like every other
  enrichment site.

Behavior-preserving: make go/test green, eslint + vitest (2270) green,
browser smoke on viewer detail, correction modal, and overlay clean.
